Address PR feedback: move payload definition to mimirpb package.
