+++ pkg/mimirpb/query_response_extra.go
@@ -1,6 +1,6 @@
 // SPDX-License-Identifier: AGPL-3.0-only
 
-package querypb
+package mimirpb
 
 import "fmt"
